requirements.txt
alembic==1.5.8
aniso8601==9.0.1
appdirs==1.4.4
autopep8==1.5.6
black==20.8b1
blinker==1.4
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
Flask-WTF==0.14.3
greenlet==1.0.0
idna==3.1
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.25.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.8.1
Pillow==8.2.0
pycodestyle==2.7.0
PyJWT==2.1.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
regex==2021.4.4
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
WTForms==2.3.3
